+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ## Version 0.9.1
 - Python build system with pyproject.toml
 - Add project to PyPi
+
+## Version 0.9.3
+- Debug fraction
+- Script for adding __insert_time__, __update_time__ (xerial-backup-enable)